+++ CHANGELOG.md
@@ -21,7 +21,11 @@ and this project adheres to
   variables).
 - Add measurement outliers detection and removal.
 - Add standard deviation to JSON report.
-- Add `samples: { warmup: number, benchmark: number }` to run options specifying
-  the number of samples to run during warmup and benchmark.
+- Add `samples: { warmup: number; benchmark: number }` to `run()` options
+  specifying the number of samples to run during warmup and benchmark.
+- Add
+  `options: { before: () => void | Promise<void>; after: () => void | Promise<void> }`
+  to `bench()` and `baseline()` specifying actions to run before and after the
+  benchmark.
 
 ## [0.1.11] - 2024-02-20